Solve accent and umlaut problem
The solution is horrible and suboptimal, but works for now. I should check how to set-up encoding for the special characters I use (Spanish and German).

A problem now is that name and surname are not separated, is this a template (tewntyseconds) problem? The same happens with Marie Curie's trial.